chore: move palette into colorschemes
