+++ frontend/src/theme/colorSchemes/dark.ts
@@ -14,50 +14,6 @@ const grey: Partial<Color> = {
   900: "#161C24",
 };
 
-export const light: PaletteOptions = {
-  primary: {
-    main: "#2D6365",
-    contrastText: "#fff",
-  },
-  secondary: {
-    main: "#F37F31",
-    contrastText: "#fff",
-  },
-  info: {
-    main: "#00B1ED",
-    contrastText: "#fff",
-  },
-  error: {
-    main: "#FF3B30",
-    contrastText: "#fff",
-  },
-  warning: {
-    main: "#FF9500",
-    contrastText: "#fff",
-  },
-  grey,
-  success: {
-    main: "#34C759",
-    contrastText: "#fff",
-  },
-  contrast: {
-    light: "#3A3A3C",
-    dark: "#1C1C1E",
-    main: "#2C2C2E",
-    contrastText: "#fff",
-    lightChannel: colorChannel("#3A3A3C"),
-    darkChannel: colorChannel("#1C1C1E"),
-    mainChannel: colorChannel("#2C2C2E"),
-    contrastTextChannel: colorChannel("#fff"),
-  },
-  background: {
-    elevated: "#f0f0f0",
-    default: "#fff",
-    paper: "#fff",
-  },
-  shadowChannel: colorChannel("#CCCCCC"),
-};
-
 export const dark: PaletteOptions = {
   background: {
     elevated: grey[900],
