+++ frontend/src/theme/colorSchemes/index.ts
@@ -1,6 +1,7 @@
 import { CssVarsThemeOptions } from "@mui/material";
 
-import { light, dark } from "../palette";
+import { dark } from "./dark";
+import { light } from "./light";
 
 export const colorSchemes: CssVarsThemeOptions["colorSchemes"] = {
   light: {
